backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==4.0.1",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "google-generativeai>=0.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.5.2",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.18",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.32.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
